--- src/neptune/common/hardware/metrics/reports/metric_reporter_factory.py
@@ -1,24 +1,0 @@
-#
-# Copyright (c) 2019, Neptune Labs Sp. z o.o.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-from neptune.common.hardware.metrics.reports.metric_reporter import MetricReporter
-
-
-class MetricReporterFactory(object):
-    def __init__(self, reference_timestamp):
-        self.__reference_timestamp = reference_timestamp
-
-    def create(self, metrics):
-        return MetricReporter(metrics=metrics, reference_timestamp=self.__reference_timestamp)
