+++ src/neptune/internal/hardware/gauges/gauge.py
