+++ src/neptune/internal/hardware/gauges/gauge_factory.py
@@ -13,16 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from neptune.common.hardware.gauges.cpu import (
+from neptune.internal.hardware.gauges.cpu import (
     CGroupCpuUsageGauge,
     SystemCpuUsageGauge,
 )
-from neptune.common.hardware.gauges.gauge_mode import GaugeMode
-from neptune.common.hardware.gauges.gpu import (
+from neptune.internal.hardware.gauges.gauge_mode import GaugeMode
+from neptune.internal.hardware.gauges.gpu import (
     GpuMemoryGauge,
     GpuUsageGauge,
 )
-from neptune.common.hardware.gauges.memory import (
+from neptune.internal.hardware.gauges.memory import (
     CGroupMemoryUsageGauge,
     SystemMemoryUsageGauge,
 )
