+++ src/neptune/internal/hardware/gauges/gauge_mode.py
