+++ src/neptune/internal/hardware/hardware_metric_reporting_job.py
@@ -24,15 +24,15 @@ from typing import (
     Optional,
 )
 
-from neptune.common.hardware.gauges.gauge_factory import GaugeFactory
-from neptune.common.hardware.gauges.gauge_mode import GaugeMode
-from neptune.common.hardware.metrics.metrics_factory import MetricsFactory
-from neptune.common.hardware.metrics.reports.metric_reporter import MetricReporter
-from neptune.common.hardware.metrics.reports.metric_reporter_factory import MetricReporterFactory
-from neptune.common.hardware.resources.system_resource_info_factory import SystemResourceInfoFactory
-from neptune.common.hardware.system.system_monitor import SystemMonitor
 from neptune.internal.background_job import BackgroundJob
+from neptune.internal.hardware.gauges.gauge_factory import GaugeFactory
+from neptune.internal.hardware.gauges.gauge_mode import GaugeMode
 from neptune.internal.hardware.gpu.gpu_monitor import GPUMonitor
+from neptune.internal.hardware.metrics.metrics_factory import MetricsFactory
+from neptune.internal.hardware.metrics.reports.metric_reporter import MetricReporter
+from neptune.internal.hardware.metrics.reports.metric_reporter_factory import MetricReporterFactory
+from neptune.internal.hardware.resources.system_resource_info_factory import SystemResourceInfoFactory
+from neptune.internal.hardware.system.system_monitor import SystemMonitor
 from neptune.internal.threading.daemon import Daemon
 from neptune.internal.utils.logger import get_logger
 from neptune.internal.utils.utils import in_docker
