+++ src/neptune/internal/hardware/resources/__init__.py
