+++ src/neptune/internal/hardware/resources/system_resource_info_factory.py
@@ -14,10 +14,10 @@
 # limitations under the License.
 #
 
-from neptune.common.hardware.cgroup.cgroup_monitor import CGroupMonitor
-from neptune.common.hardware.gauges.gauge_mode import GaugeMode
-from neptune.common.hardware.resources.gpu_card_indices_provider import GPUCardIndicesProvider
-from neptune.common.hardware.resources.system_resource_info import SystemResourceInfo
+from neptune.internal.hardware.cgroup.cgroup_monitor import CGroupMonitor
+from neptune.internal.hardware.gauges.gauge_mode import GaugeMode
+from neptune.internal.hardware.resources.gpu_card_indices_provider import GPUCardIndicesProvider
+from neptune.internal.hardware.resources.system_resource_info import SystemResourceInfo
 
 
 class SystemResourceInfoFactory(object):
